Update plist files for server versions before 14
Make pkg-plist-server for versions before 14 conform to the same
conventions previously applied to the the plist file for version 14.

Add an experimental plist file that contains selectors to exclude
files not in version 13 or 14 (needs not yet implemented Makefile
support). This adds too much complexity to be practical for a port
converted to use FLAVORs, but I include it as an example.